converting formstate to redux. created slice. outsourced formdata interface
