+++ src/models/FormData.ts
@@ -1,0 +1,10 @@
+export interface FormData {
+     name?: string;
+  surname?: string;
+    email?: string;
+    phone?: string;
+     from?: string;
+     till?: string;
+   pickup?: string;
+  dropoff?: string;
+}
