+++ src/util/validateForm.ts
@@ -1,15 +1,5 @@
 import User from '@/models/User';
-
-export interface FormData {
-     name?: string;
-  surname?: string;
-    email?: string;
-    phone?: string;
-     from?: string;
-     till?: string;
-   pickup?: string;
-  dropoff?: string;
-}
+import { FormData } from '@/models/FormData'
 
 function validateEmptyFields(data: FormData) {
   const errors: Partial<FormData> = {};
